Changed ReprojectGeometry to accept either CRS struct or a StringType.
