--- pyrasterframes/src/main/python/docs/languages.md
@@ -1,2 +1,0 @@
-# API Languages
-
