+++ pyrasterframes/src/main/python/tests/GeoTiffWriterTests.py
@@ -61,3 +61,4 @@ class GeoTiffWriter(TestEnvironment):
             self.assertEqual((f.width, f.height), (128, 128))
 
         os.remove(dest)
+
